fix test for as_array
1. start from 2 AcquisitionData created from a AcquisitionGeometry, with
different values inside.
2. split the 2 AcquisitionData in equivalent subsets
3. the test id(data1.as_array()) != id(data2.as_array()) failed in python 3.5 and numpy 1.11

In numpy 1.11 the as_array() method with subsets, will/may return an object on the same
memory location. I suppose this is just because the data is allocated, the pointer is
returned and then the whole thing is discarded. The subsequent call allocates the same
memory location and the test fails.
The fix comes by storing d1 = data1.as_array() and d2.as_array() and testing that
 id(d1) != id(d2), which works