Fix argument order for subtrack and divide between different container types
Closes #2130

When doing an operation the __container_priority__ is compared with other. If other has a higher priority, then the operator is called from other with self as the argument. This is fine for commutative operations, but not for sub and divide.

Use -other + self and other^-1 * self to ensure correct result